requirements.txt
pyspark
kafka-python
pandas
matplotlib
psutil
requests
jupyter